[cleaned] the big mess of ints vs uints vs bytes [added] vector3b Note : be careful, when you depend on y == -1 to exit loops, y cant be a byte !

+++ TechCraft/NewTake/NewTake/model/World.cs
@@ -85,22 +85,21 @@ namespace NewTake.model
             return setBlock(pos.X,pos.Y,pos.Z,b);
         }
 
-        //this method is only invoked by user action, not by the engine so optimisation is not a problem for now
         public Block setBlock(uint x, uint y, uint z, Block newType)
         {
             if (InView(x, y, z))
             {
                 Chunk chunk = viewableChunks[x / Chunk.CHUNK_XMAX, z / Chunk.CHUNK_ZMAX];
-                
-                uint localX = x % Chunk.CHUNK_XMAX;
-                uint localY = y % Chunk.CHUNK_YMAX;
-                uint localZ = z % Chunk.CHUNK_ZMAX;
+
+                byte localX = (byte)(x % Chunk.CHUNK_XMAX);
+                byte localY = (byte)(y % Chunk.CHUNK_YMAX);
+                byte localZ = (byte)(z % Chunk.CHUNK_ZMAX);
                 //TODO messy chunk coordinates types
                 //Block old = chunk.Blocks[localX, localY, localZ];
                 Block old = chunk.Blocks[localX * Chunk.FlattenOffset + localZ * Chunk.CHUNK_YMAX + localY];
                 
                //chunk.setBlock is also called by terrain generators for Y loops min max optimisation
-               chunk.setBlock((int)localX, (int)localY, (int)localZ, new Block(newType.Type, old.LightAmount));
+               chunk.setBlock(localX, localY, localZ, new Block(newType.Type, old.LightAmount));
                //TODO ( maybe ? )  when digging, mark neighbours chunks as dirty to fill rendering holes                                       
                
                 chunk.dirty = true;
